Extract scan and firewall headers
