+++ projects/plugins/protect/src/js/components/error-admin-section-hero/styles.module.scss
