--- projects/plugins/protect/src/js/components/scanning-header/styles.module.scss
@@ -1,10 +1,0 @@
-.illustration {
-	display: flex;
-	align-items: center;
-	height: 100%;
-	img {
-		// let's fit images
-		object-fit: cover;
-		width: 100%;
-	}
-}
