+++ projects/plugins/protect/src/js/routes/scan/index.jsx
@@ -1,29 +1,14 @@
-import {
-	AdminSection,
-	Container,
-	Col,
-	Text,
-	useBreakpointMatch,
-	Status,
-} from '@automattic/jetpack-components';
-import { dateI18n } from '@wordpress/date';
-import { __, _n, sprintf } from '@wordpress/i18n';
-import { useMemo, useState } from 'react';
+import { AdminSection, Container, Col } from '@automattic/jetpack-components';
 import AdminPage from '../../components/admin-page';
-import AdminSectionHero from '../../components/admin-section-hero';
-import ErrorHeader from '../../components/error-header';
-import OnboardingPopover from '../../components/onboarding-popover';
 import ScanFooter from '../../components/scan-footer';
-import ScanNavigation from '../../components/scan-navigation';
-import ScanningHeader from '../../components/scanning-header';
 import ThreatsList from '../../components/threats-list';
-import useScanStatusQuery, { isScanInProgress } from '../../data/scan/use-scan-status-query';
+import useScanStatusQuery from '../../data/scan/use-scan-status-query';
 import useAnalyticsTracks from '../../hooks/use-analytics-tracks';
 import { OnboardingContext } from '../../hooks/use-onboarding';
 import usePlan from '../../hooks/use-plan';
 import useProtectData from '../../hooks/use-protect-data';
 import onboardingSteps from './onboarding-steps';
-import styles from './styles.module.scss';
+import ScanAdminSectionHero from './scan-admin-section-hero';
 
 /**
  * Scan Page
@@ -41,10 +26,6 @@ const ScanPage = () => {
 		lastChecked,
 	} = useProtectData();
 	const { data: status } = useScanStatusQuery( { usePolling: true } );
-	const [ isSm ] = useBreakpointMatch( 'sm' );
-
-	// Popover anchor
-	const [ dailyScansPopoverAnchor, setDailyScansPopoverAnchor ] = useState( null );
 
 	let currentScanStatus;
 	if ( status.error ) {
@@ -64,89 +45,10 @@ const ScanPage = () => {
 		},
 	} );
 
-	let lastCheckedLocalTimestamp = null;
-	if ( lastChecked ) {
-		// Convert the lastChecked UTC date to a local timestamp
-		lastCheckedLocalTimestamp = new Date( lastChecked + ' UTC' ).getTime();
-	}
-
-	const renderHeader = useMemo( () => {
-		if ( isScanInProgress( status ) ) {
-			return <ScanningHeader currentProgress={ status?.currentProgress } />;
-		}
-
-		if ( status.error ) {
-			return (
-				<ErrorHeader
-					baseErrorMessage={ __( 'We are having problems scanning your site.', 'jetpack-protect' ) }
-					errorMessage={ status.errorMessage }
-					errorCode={ status.errorCode }
-				/>
-			);
-		}
-
-		return (
-			<AdminSectionHero
-				main={
-					<>
-						<Status status={ 'active' } label={ __( 'Active', 'jetpack-protect' ) } />
-						<AdminSectionHero.Heading showIcon>
-							{ numThreats > 0
-								? sprintf(
-										/* translators: %s: Total number of threats/vulnerabilities */
-										__( '%1$s %2$s found', 'jetpack-protect' ),
-										numThreats,
-										hasPlan
-											? _n( 'threat', 'threats', numThreats, 'jetpack-protect' )
-											: _n( 'vulnerability', 'vulnerabilities', numThreats, 'jetpack-protect' )
-								  )
-								: sprintf(
-										/* translators: %s: Pluralized type of threat/vulnerability */
-										__( 'No %s found', 'jetpack-protect' ),
-										hasPlan
-											? __( 'threats', 'jetpack-protect' )
-											: __(
-													'vulnerabilities',
-													'jetpack-protect',
-													/* dummy arg to avoid bad minification */ 0
-											  )
-								  ) }
-						</AdminSectionHero.Heading>
-						<AdminSectionHero.Subheading>
-							<>
-								<Text ref={ setDailyScansPopoverAnchor }>
-									{ lastCheckedLocalTimestamp ? (
-										<>
-											<span className={ styles[ 'subheading-content' ] }>
-												{ dateI18n( 'F jS g:i A', lastCheckedLocalTimestamp ) }
-											</span>
-											&nbsp;
-											{ __( 'results', 'jetpack-protect' ) }
-										</>
-									) : (
-										__( 'Most recent results', 'jetpack-protect' )
-									) }
-								</Text>
-								{ ! hasPlan && (
-									<OnboardingPopover
-										id="free-daily-scans"
-										position={ isSm ? 'bottom' : 'middle left' }
-										anchor={ dailyScansPopoverAnchor }
-									/>
-								) }
-							</>
-						</AdminSectionHero.Subheading>
-						<ScanNavigation />
-					</>
-				}
-			/>
-		);
-	}, [ status, lastCheckedLocalTimestamp, numThreats, hasPlan, isSm, dailyScansPopoverAnchor ] );
-
 	return (
 		<OnboardingContext.Provider value={ onboardingSteps }>
 			<AdminPage>
-				{ renderHeader }
+				<ScanAdminSectionHero />
 				{ ( ! status.error || ( status.error && numThreats > 0 ) ) && (
 					<AdminSection>
 						{ ' ' }
